Fixdev

+++ src/main/java/tdc/edu/vn/project_mobile_be/dtos/requests/product/ProductCreateRequestDTO.java
@@ -30,7 +30,7 @@ public class ProductCreateRequestDTO implements IDto<Product> {
 
     @Min(value = 0, message = "ProductPrice phải lớn hơn hoặc bằng 0")
     @JsonProperty("productPrice")
-    private BigDecimal productPrice;
+    private double productPrice;
 
     @Min(value = 0, message = "ProductQuantity phải lớn hơn hoặc bằng 0")
     @JsonProperty("productQuantity")
